Word Datei Erstellung in eigene Klasse ausgelagert
